+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add handling to check parents when innerText is blank
+- Add an observer to capture DOM elements added after the Interactor is initialized
 
 ### Fixed